+++ src/main/java/com/kuit/findyou/domain/report/dto/request/RetrieveReportRequestDTO.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.report.dto;
+package com.kuit.findyou.domain.report.dto.request;
 
 import io.swagger.v3.oas.annotations.Parameter;
 import jakarta.validation.constraints.NotNull;
@@ -6,14 +6,13 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import java.time.LocalDate;
 
 @Getter
 @Setter
 @ParameterObject
-public class RetrieveReportRequest {
+public class RetrieveReportRequestDTO {
 
     @NotNull
     @Parameter(description = "이전 요청을 통해 받아온 데이터들 중 마지막 신고글의 ID 입니다. 이 값을 다음 요청에 포함시키면 그 다음 신고글들을 조회하여 응답합니다.")
