+++ src/main/java/com/kuit/findyou/domain/report/dto/request/WitnessReportDTO.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.report.dto;
+package com.kuit.findyou.domain.report.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
@@ -9,23 +9,23 @@ import java.util.List;
 
 @Getter
 @Builder
-public class MissingReportDTO {
+public class WitnessReportDTO {
     @Schema(description = "업로드한 이미지 url 리스트")
     private List<String> imageUrls;
     @Schema(description = "품종", example = "말티즈")
     private Long breed;
     @Schema(description = "성별", example = "M")
     private String sex;
-    @Schema(description = "털색", example = "흰색")
+    @Schema(description = "털색")
     private List<String> furColor;
-    @Schema(description = "실종 장소")
+    @Schema(description = "목격 장소")
     private String location;
     @Schema(description = "특징 ID 리스트")
     private List<Long> features;
     @Schema(description = "추가 설명")
     private String description;
-    @Schema(description = "실종 날짜" , example = "2025-02-07")
-    private LocalDate missingDate;
+    @Schema(description = "목격 날짜" , example = "2025-02-07")
+    private LocalDate foundDate;
     @Schema(description = "유저 ID")
     private Long userId;
 }
