+++ src/main/java/com/kuit/findyou/domain/report/service/ReportAnimalInfoService.java
@@ -2,7 +2,7 @@ package com.kuit.findyou.domain.report.service;
 
 import com.kuit.findyou.domain.user.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
-import com.kuit.findyou.domain.report.dto.ReportInfoDTO;
+import com.kuit.findyou.domain.report.dto.response.ReportInfoDTO;
 import com.kuit.findyou.domain.report.model.*;
 import com.kuit.findyou.domain.report.repository.*;
 import com.kuit.findyou.global.common.exception.ReportNotFoundException;
