+++ src/main/java/com/kuit/findyou/domain/report/service/ReportAnimalRetrieveService.java
@@ -2,8 +2,8 @@ package com.kuit.findyou.domain.report.service;
 
 import com.kuit.findyou.domain.user.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
-import com.kuit.findyou.domain.report.dto.Card;
-import com.kuit.findyou.domain.report.dto.ReportCardDTO;
+import com.kuit.findyou.domain.report.dto.response.Card;
+import com.kuit.findyou.domain.report.dto.response.ReportCardDTO;
 import com.kuit.findyou.domain.report.model.Report;
 import com.kuit.findyou.domain.report.repository.ReportRepository;
 import com.kuit.findyou.global.common.exception.UserNotFoundException;
