+++ src/main/java/com/kuit/findyou/domain/user/dto/request/PostInterestAnimalRequest.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.user.dto;
+package com.kuit.findyou.domain.user.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
