+++ src/main/java/com/kuit/findyou/domain/user/service/InterestAnimalRetrieveService.java
@@ -5,8 +5,8 @@ import com.kuit.findyou.domain.report.model.InterestProtectingReport;
 import com.kuit.findyou.domain.report.model.InterestReport;
 import com.kuit.findyou.domain.report.repository.InterestProtectingReportRepository;
 import com.kuit.findyou.domain.report.repository.InterestReportRepository;
-import com.kuit.findyou.domain.user.dto.GetInterestAnimalCursorPageDto;
-import com.kuit.findyou.domain.user.dto.InterestAnimalCard;
+import com.kuit.findyou.domain.user.dto.response.GetInterestAnimalCursorPageDto;
+import com.kuit.findyou.domain.user.dto.response.InterestAnimalCard;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
