+++ src/test/java/com/kuit/findyou/domain/report/service/AnimalRetrieveServiceTest.java
@@ -7,8 +7,8 @@ import com.kuit.findyou.domain.auth.repository.UserRepository;
 import com.kuit.findyou.domain.home.dto.ReportTag;
 import com.kuit.findyou.domain.image.model.Image;
 import com.kuit.findyou.domain.image.repository.ImageRepository;
-import com.kuit.findyou.domain.report.dto.Card;
-import com.kuit.findyou.domain.report.dto.TotalCardDTO;
+import com.kuit.findyou.domain.report.dto.response.Card;
+import com.kuit.findyou.domain.report.dto.response.TotalCardDTO;
 import com.kuit.findyou.domain.report.model.*;
 import com.kuit.findyou.domain.report.repository.*;
 import lombok.extern.slf4j.Slf4j;
