+++ src/test/java/com/kuit/findyou/domain/report/service/InterestAnimalRetriveServiceTest.java
@@ -7,8 +7,8 @@ import com.kuit.findyou.domain.report.model.ProtectingReport;
 import com.kuit.findyou.domain.report.model.Sex;
 import com.kuit.findyou.domain.report.repository.InterestProtectingReportRepository;
 import com.kuit.findyou.domain.report.repository.ProtectingReportRepository;
-import com.kuit.findyou.domain.user.dto.GetInterestAnimalCursorPageDto;
-import com.kuit.findyou.domain.user.dto.PostInterestAnimalRequest;
+import com.kuit.findyou.domain.user.dto.response.GetInterestAnimalCursorPageDto;
+import com.kuit.findyou.domain.user.dto.request.PostInterestAnimalRequest;
 import com.kuit.findyou.domain.user.service.InterestAnimalRetrieveService;
 import com.kuit.findyou.domain.user.service.UserService;
 import jakarta.persistence.EntityManager;
