+++ src/test/java/com/kuit/findyou/domain/report/service/ProtectingAnimalInfoServiceTest.java
@@ -2,7 +2,7 @@ package com.kuit.findyou.domain.report.service;
 
 import com.kuit.findyou.domain.user.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
-import com.kuit.findyou.domain.report.dto.ProtectingReportInfoDTO;
+import com.kuit.findyou.domain.report.dto.response.ProtectingReportInfoDTO;
 import com.kuit.findyou.domain.report.model.*;
 import com.kuit.findyou.domain.report.repository.InterestProtectingReportRepository;
 import com.kuit.findyou.domain.report.repository.ProtectingReportRepository;
