+++ src/test/java/com/kuit/findyou/domain/report/service/ProtectingAnimalRetrieveServiceTest.java
@@ -2,8 +2,8 @@ package com.kuit.findyou.domain.report.service;
 
 import com.kuit.findyou.domain.user.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
-import com.kuit.findyou.domain.report.dto.Card;
-import com.kuit.findyou.domain.report.dto.ProtectingReportCardDTO;
+import com.kuit.findyou.domain.report.dto.response.Card;
+import com.kuit.findyou.domain.report.dto.response.ProtectingReportCardDTO;
 import com.kuit.findyou.domain.report.model.InterestProtectingReport;
 import com.kuit.findyou.domain.report.model.Neutering;
 import com.kuit.findyou.domain.report.model.ProtectingReport;
