+++ src/test/java/com/kuit/findyou/domain/user/service/UserServiceTest.java
@@ -8,8 +8,8 @@ import com.kuit.findyou.domain.home.dto.ReportTag;
 import com.kuit.findyou.domain.image.model.Image;
 import com.kuit.findyou.domain.report.model.*;
 import com.kuit.findyou.domain.report.repository.*;
-import com.kuit.findyou.domain.user.dto.GetUsersReportsResponse;
-import com.kuit.findyou.domain.user.dto.PostInterestAnimalRequest;
+import com.kuit.findyou.domain.user.dto.response.GetUsersReportsResponse;
+import com.kuit.findyou.domain.user.dto.request.PostInterestAnimalRequest;
 import com.kuit.findyou.domain.user.exception.AlreadySavedInterestException;
 import com.kuit.findyou.global.common.exception.ReportNotFoundException;
 import com.kuit.findyou.global.common.exception.UserNotFoundException;
