+++ src/test/java/com/kuit/findyou/domain/user/service/ViewedAnimalRetrieveServiceTest.java
@@ -7,10 +7,10 @@ import com.kuit.findyou.domain.auth.repository.UserRepository;
 import com.kuit.findyou.domain.home.dto.ReportTag;
 import com.kuit.findyou.domain.image.model.Image;
 import com.kuit.findyou.domain.image.repository.ImageRepository;
-import com.kuit.findyou.domain.report.dto.Card;
+import com.kuit.findyou.domain.report.dto.response.Card;
 import com.kuit.findyou.domain.report.model.*;
 import com.kuit.findyou.domain.report.repository.*;
-import com.kuit.findyou.domain.user.dto.ViewedCardDTO;
+import com.kuit.findyou.domain.user.dto.response.ViewedCardDTO;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
